+++ __lstat64_time64.c
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Gaël PORTAY
+ * Copyright 2023-2024 Gaël PORTAY
  *
  * SPDX-License-Identifier: LGPL-2.1-or-later
  */
@@ -7,7 +7,10 @@
 #ifdef __linux__
 #include <stdio.h>
 #include <errno.h>
+#include <limits.h>
+#include <dlfcn.h>
 #include <fcntl.h>
+#include <sys/xattr.h>
 
 #include <sys/stat.h>
 
@@ -16,15 +19,52 @@
 #ifdef __GLIBC__
 #ifdef _LARGEFILE64_SOURCE
 #if __TIMESIZE == 32
-extern int __fstatat64_time64(int, const char *, struct stat64 *, int);
+static int (*sym)(const char *, struct stat64 *);
+
+hidden int next___lstat64_time64(const char *path, struct stat64 *statbuf)
+{
+	if (!sym)
+		sym = dlsym(RTLD_NEXT, "__lstat64_time64");
+
+	if (!sym)
+		return __dl_set_errno_and_perror(ENOSYS, -1);
+
+	return sym(path, statbuf);
+}
 
 int __lstat64_time64(const char *path, struct stat64 *statbuf)
 {
-	__debug("%s(path: '%s', ...)\n", __func__, path);
+	char buf[PATH_MAX];
+	int ret = -1;
+	ssize_t siz;
+	uid_t uid;
+	gid_t gid;
 
-	/* Forward to another function */
-	return __fstatat64_time64(AT_FDCWD, path, statbuf,
-				  AT_SYMLINK_NOFOLLOW);
+	siz = path_resolution(AT_FDCWD, path, buf, sizeof(buf),
+			      AT_SYMLINK_NOFOLLOW);
+	if (siz == -1)
+		goto exit;
+
+	ret = next___lstat64_time64(buf, statbuf);
+	if (ret == -1)
+		goto exit;
+
+	uid = __get_uid(buf);
+	if (uid == (uid_t)-1)
+		statbuf->st_uid = 0;
+
+	gid = __get_gid(buf);
+	if (gid == (gid_t)-1)
+		statbuf->st_gid = 0;
+
+	__st_mode(buf, statbuf);
+	__st_uid(buf, statbuf);
+	__st_gid(buf, statbuf);
+
+exit:
+	__debug("%s(path: '%s' -> '%s', ...) -> %i\n", __func__, path, buf, ret);
+
+	return ret;
 }
 #endif
 #endif